Network data on comments

This dataset includes network data on comments. Network data is similar to directed dyad-year data except that it only includes directed dyad-years with at least one comment. For every year, there is one node per member state or institution that can submit a comment or notify a proposed technical regulation (only member states). Links can only exist between a third-party member state or institution that can submit a comment and a member state that can notify a technical regulation. There is a link between a third-party member state or institution and a notifying member state if and only if the former submitted at least one comment with respect to a technical regulation notified by the latter. The weight of the edge is the number of comments submitted. There is one observation per third-party member state or institution per notifying member state per year (2002-2020), excluding directed dyad-years where a state was not a member of the Single Market, but only if count of comments is positive.

comments_net

Format

A data frame with 7 variables:

key_id

Numeric. An ID number that uniquely identifies each observation in the dataset.

time

Numeric. The time period, where 1988 is coded 1.

from_node_id

Numeric. An ID number that uniquely identifies each node (a member state or institution) that creates a link.

from_node

String. The name of the member state or institution that submitted the opinion to the Commission.

to_node_id

Numeric. An ID number that uniquely identifies each node (a member state) that receives a link.

to_node

String. The name of the member state that notified the technical regulation to the Commission.

edge_weight

Numeric. The weight of the edge, which is the number of comments filed against the member state the notified the proposed technical regulation.