Data on notifications

This dataset include data on technical regulations proposed at the national level by member states of the EU Single Market (including members of the Single Market that are not members of the EU) that are notified to the Commission under the 2015/1535 procedure. There is one observation per notified technical regulation (1988-2020). The dataset includes information on the start and end dates of the case, the member state that notified the proposed technical regulation, which third-party member states or institutions (if any) submitted comments and/or detailed opinions, and whether the Commission submitted a comment and/or detailed opinion.

notifications

Format

A data frame with 16 variables:

key_id

Numeric. An ID number that uniquely identifies each observation in the dataset.

notification_id

String. An ID number that uniquely identifies the notification in the format TRIS:####:####:XX:N, where the first set of digits is the year of the notification, the second set of digits is the notification number, and XX is the two-digit code assigned by the Commission to the member state that notified the proposed technical regulation.

notification_by_id

Numeric. An ID number that uniquely identifies the member state that notified the proposed technical regulation to the Commission.

notification_by

String. The name of the member state that notified the proposed technical regulation to the Commission.

notification_by_code

String. A multi-letter code assigned by the Commission that uniquely identifies the member state that notified the proposed technical regulation to the Commission.

start_date

String. The date that the Commission opened the case.

start_year

Numeric. The year that the Commission opened the case.

end_date

String. The date that the Commission closed the case.

end_year

Numeric. The year that the Commission closed the case.

postponement

Dummy. A dummy variable indicating whether the notified technical regulation was postponed. Coded 1 if the technical regulation was postponed and 0 otherwise.

comments

String. A list of the member states and institutions that submitted comments with respect to the proposed technical regulation, separated by commas. Coded NA if no member states or institutions submitted opinions.

count_comments

Numeric. A count of the number of comments submitted by the third-party member state or institution.

commission_comment

Dummy. A dummy variable indicating whether the Commission submitted a comment with respect to the notified proposed technical regulation. Coded 1 if the Commission submitted a comment and 0 otherwise.

opinions

String. A list of the member states and institutions that submitted opinions with respect to the proposed technical regulation, separated by commas. Coded NA if no member states or institutions submitted opinions.

count_opinions

Numeric. A count of the number of opinions submitted by the third-party member state or institution.

commission_opinion

Dummy. A dummy variable indicating whether the Commission submitted an opinion with respect to the notified proposed technical regulation. Coded 1 if the Commission submitted an opinion and 0 otherwise.