Data on detailed opinions

This dataset includes data on detailed opinions submitted by third-party member states and institutions in reponse to proposed technical regulations notified by member states (1988-2020). There is one observation per detailed opinion. The dataset indicates the the member state or institution that submitted the detailed opinion and the member state that notified the proposed technical regulation to the Commission.

opinions

Format

A data frame with 13 variables:

key_id

Numeric. An ID number that uniquely identifies each observation in the dataset.

notification_id

String. An ID number that uniquely identifies the notification in the format TRIS:####:####:XX:N, where the first set of digits is the year of the notification, the second set of digits is the notification number, and XX is the two-digit code assigned by the Commission to the member state that notified the proposed technical regulation.

notification_by_id

Numeric. An ID number that uniquely identifies the member state that notified the proposed technical regulation to the Commission.

notification_by

String. The name of the member state that notified the proposed technical regulation to the Commission.

notification_by_code

String. A multi-letter code assigned by the Commission that uniquely identifies the member state that notified the proposed technical regulation to the Commission.

start_date

String. The date that the Commission opened the case.

start_year

Numeric. The year that the Commission opened the case.

end_date

String. The date that the Commission closed the case.

end_year

Numeric. The year that the Commission closed the case.

opinion_id

String. An ID number that uniquely identifies the opinion in the format TRIS:####:####:XX:O:YY, where the first set of digits is the year of the notification, the second set of digits is the notification number, XX is the two-letter code assigned by the Commission to the member state that notified the proposed technical regulation, and YY is the multi-letter code assigned by the Commission to the member state or institution that submitted the opinion.

opinion_by_id

Numeric. An ID number that uniquely identifies the member state or institution that submitted the detailed opinion.

opinion_by

String. The name of the member state or institution that submitted the detailed opinion.

opinion_by_code

String. A multi-letter code assigned by the Commission that uniquely identifies the member state or institution that submitted the detailed opinion.